Refactor and Clean Up Duplicate Code for Embedding Providers
